+++ webknossos/webknossos/annotation/_nml_conversion.py
@@ -1,6 +1,7 @@
 import colorsys
 import itertools
 import warnings
+from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
 import networkx as nx
@@ -40,11 +41,17 @@ def nml_to_skeleton(nml: wknml.Nml) -> "Skeleton":
     for nml_tree in nml.trees:
         if nml_tree.groupId is None:
             new_tree = skeleton.add_tree(
-                nml_tree.name, _enforced_id=nml_tree.id, color=nml_tree.color
+                nml_tree.name,
+                _enforced_id=nml_tree.id,
+                color=nml_tree.color,
+                metadata={i.key: i.value for i in nml_tree.metadata},
             )
         else:
             new_tree = groups_by_id[nml_tree.groupId].add_tree(
-                nml_tree.name, _enforced_id=nml_tree.id, color=nml_tree.color
+                nml_tree.name,
+                _enforced_id=nml_tree.id,
+                color=nml_tree.color,
+                metadata={i.key: i.value for i in nml_tree.metadata},
             )
         _nml_tree_to_wk_tree(new_tree, nml_tree)
 
@@ -68,6 +75,29 @@ def nml_to_skeleton(nml: wknml.Nml) -> "Skeleton":
         )
 
     return skeleton
+
+
+def _dict_to_metadata_entry_list(
+    metadata: dict[str, str | int | float | Sequence[str]],
+) -> list[wknml.MetadataEntry]:
+    metadata_entry_list = []
+    if metadata is not None:
+        for key, value in metadata.items():
+            if isinstance(value, str):
+                metadata_entry_list.append(
+                    wknml.MetadataEntry(key=key, type="str", value=value)
+                )
+            elif isinstance(value, int) or isinstance(value, float):
+                metadata_entry_list.append(
+                    wknml.MetadataEntry(key=key, type="number", value=value)
+                )
+            elif isinstance(value, Sequence):
+                metadata_entry_list.append(
+                    wknml.MetadataEntry(key=key, type="list", value=value)
+                )
+            else:
+                raise ValueError(f"Unsupported metadata type: {type(value)}")
+    return metadata_entry_list
 
 
 def _nml_tree_to_wk_tree(
@@ -155,6 +185,7 @@ def annotation_to_nml(
         nodes, edges = _extract_nodes_and_edges_from_tree(tree)
         color = tree.color or _random_color_rgba()
         name = tree.name or f"tree{tree.id}"
+        metadata = _dict_to_metadata_entry_list(tree.metadata) or []
 
         trees.append(
             wknml.Tree(
@@ -164,6 +195,7 @@ def annotation_to_nml(
                 name=name,
                 groupId=tree.group.id if tree.group != annotation.skeleton else None,
                 color=color,
+                metadata=metadata,
             )
         )
 
@@ -181,6 +213,7 @@ def annotation_to_nml(
                     name=segment_info.name,
                     anchor_position=segment_info.anchor_position,
                     color=segment_info.color,
+                    metadata=_dict_to_metadata_entry_list(segment_info.metadata),
                 )
                 for segment_id, segment_info in volume.segments.items()
             ],