+++ webknossos/webknossos/client/_upload_dataset.py
@@ -133,6 +133,7 @@ def upload_dataset(
             new_dataset_name,
             context.organization_id,
             total_file_count=len(file_infos),
+            total_file_size_in_bytes=total_file_size,
             layers_to_link=[
                 layer.as_api_linked_layer_identifier() for layer in layers_to_link
             ],