+++ webknossos/Changelog.md
@@ -16,6 +16,9 @@ For upgrade instructions, please check the respective _Breaking Changes_ section
 - Dropped support for Python 3.9. [#1277](https://github.com/scalableminds/webknossos-libs/pull/1277)
 
 ### Added
+- Adds the field `totalFileSizeInBytes` to the dataset upload information. [#1264](https://github.com/scalableminds/webknossos-libs/pull/1264)
+- Added support for metadata in segment and trees. [#1271](https://github.com/scalableminds/webknossos-libs/pull/1271)
+- Added a warning on import if a newer version of the `webknossos` package is available from PyPI. [#1280](https://github.com/scalableminds/webknossos-libs/pull/1280)
 
 ### Changed
 - Upgraded `tensorstore` to `0.1.72`. [#1277](https://github.com/scalableminds/webknossos-libs/pull/1277)